add mvi, changes in main screen

+++ app/src/main/java/greenfieldxd/noteable/presentation/screens/main/MainViewModel.kt
@@ -4,7 +4,9 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
 import greenfieldxd.noteable.data.repository.NoteRepository
-import greenfieldxd.noteable.domain.model.Note
+import kotlinx.coroutines.flow.SharingStarted
+import kotlinx.coroutines.flow.map
+import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
@@ -12,11 +14,40 @@ import javax.inject.Inject
 class MainViewModel @Inject constructor(
     private val noteRepository: NoteRepository
 ) : ViewModel() {
-    val allNotes = noteRepository.getAllNotes()
+    val screenState = noteRepository.getAllNotes()
+        .map { notes ->
+            if (notes.isEmpty()) MainScreenState.Empty
+            else MainScreenState.Content(notes)
+        }
+        .stateIn(
+            scope = viewModelScope,
+            started = SharingStarted.WhileSubscribed(5000),
+            initialValue = MainScreenState.Empty
+        )
 
-    fun deleteNote(note: Note) {
-        viewModelScope.launch {
-            noteRepository.deleteNote(note)
+    fun dispatch(action: MainScreenAction) {
+        when (action) {
+            is MainScreenAction.Delete -> {
+                (screenState.value as? MainScreenState.Content)?.let { contentState ->
+                    val noteToDelete = contentState.notes.find { it.id == action.id }
+                    noteToDelete?.let { note ->
+                        viewModelScope.launch {
+                            noteRepository.deleteNote(note)
+                        }
+                    }
+                }
+            }
+            is MainScreenAction.Pin -> {
+                (screenState.value as? MainScreenState.Content)?.let { contentState ->
+                    val noteToPin = contentState.notes.find { it.id == action.id }
+                    noteToPin?.let { note ->
+                        val updatedNote = note.copy(pinned = action.value)
+                        viewModelScope.launch {
+                            noteRepository.updateNote(updatedNote)
+                        }
+                    }
+                }
+            }
         }
     }
 }
